chore: Move organizations picker to ladning page sidebar
